added get notivications from account service
